Partially restructured the interaction handler, added some generic physics functions

+++ src/main.cpp
@@ -17,7 +17,7 @@ int main() {
   //initialize player
   //player pos should be determined by level file
   mapData.player.setPos(32,32);
-  mapData.player.setSize(sf::Vector2i(16,16));
+  mapData.player.setSize(sf::Vector2f(16,16));
   mapData.player.setSpeed(4);
   mapData.player.setHealth(5);
   mapData.player.setMaxCooldown(15);
